requirements.txt
acvl_utils==0.2.5
adan_pytorch==0.1.0
batchgenerators==0.25.1
batchgeneratorsv2==0.3.0
boto3==1.40.5
botocore==1.40.18
concurrentbuffer==0.0.8
connected_components_3d==3.24.0
creationism==0.0.5
dynamic_network_architectures==0.3.1
einops==0.8.1
hiddenlayer==0.3
huggingface_hub==0.29.2
huggingface_hub==0.34.4
joblib==1.4.2
jsonschema==4.25.1
lmdb==1.7.3
loguru==0.7.2
matplotlib==3.10.7
monai==1.5.1
nibabel==5.3.2
numexpr==2.12.1
numpy==2.3.4
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.3.3
Pillow==12.0.0
pytz==2024.1
pytz==2025.2
PyYAML==6.0.1
PyYAML==6.0.2
PyYAML==6.0.3
Requests==2.32.5
rtree==1.4.1
scikit_learn==1.4.2
scipy==1.16.2
seaborn==0.13.2
setuptools==65.5.0
setuptools==80.9.0
Shapely==2.1.2
SimpleITK==2.4.0
SimpleITK==2.5.2
skimage==0.0
threadpoolctl==3.6.0
tiatoolbox==1.6.0
tifffile==2024.9.20
timm==1.0.20
# torch==2.3.0
# torch==2.8.0+cu129
torchmetrics==1.8.2
torchstain==1.4.1
# torchvision==0.23.0+cu129
tqdm==4.66.4
tqdm==4.67.1
xgboost==3.1.0
